requirements.txt
fastapi
uvicorn
pydantic
openai
langchain-community
langchain-text-splitters
pypdf
chromadb
httpx
python-multipart
docker
load_dotenv